refactor(sync): extract toBase64/fromBase64 to base64Utils.ts
Part of Issue #180 structural refactor. Extracts UTF-8 safe
base64 encode/decode utilities to a dedicated module.

Fixes #180
